adding gRPC support

+++ SmartStore.Application/Features/Products/Commands/CreateProductCommand.cs
@@ -2,5 +2,5 @@
 
 namespace SmartStore.Application.Features.Products.Commands
 {
-    public record CreateProductCommand(string Name, decimal Price) : IRequest<Guid>;
+    public record CreateProductCommand(string Name, double Price) : IRequest<Guid>;
 }
