+++ SmartStore.Domain/Entities/Product.cs
@@ -7,17 +7,17 @@
             
         }
 
-        public Product(Guid id, string name, decimal price, bool? isAvailable = true)
+        public Product(Guid id, string name, double price, bool? isAvailable = true)
         {
             InitializeProduct(id, name, price, isAvailable);
         }
 
-        public Product(string name, decimal price, bool? isAvailable = true)
+        public Product(string name, double price, bool? isAvailable = true)
         {
             InitializeProduct(Guid.NewGuid(), name, price, isAvailable);
         }
 
-        private void InitializeProduct(Guid id, string name, decimal price, bool? isAvailable)
+        private void InitializeProduct(Guid id, string name, double price, bool? isAvailable)
         {
             Id = id;
             Name = name;
@@ -27,13 +27,13 @@
         }
 
         public string Name { get; private set; } = string.Empty;
-        public decimal Price { get; private set; }
+        public double Price { get; private set; }
         public bool IsAvailable { get; private set; }
         public int SalesCount { get; private set; }
 
         public void MarkAsSoldOut() => IsAvailable = false;
         public void AddSeles(int sales) => SalesCount = SalesCount + sales;
         public void SetName(string name) => Name = name;
-        public void SetPrice(decimal price) => Price = price;
+        public void SetPrice(double price) => Price = price;
     }
 }
